Add typing to `BaseRunner.__init__`
The `BaseRunner` class is a widely inherited class, and it's untyped.
Sometimes the wrong thing gets passed into the classes that inherit
`BaseRunner`. The "subclasses" of `BaseRunner` generally further
restrict some of the available typing, (f.x. the type of node that
can be passed in). In order to properly type those subclasses, we
first should type `BaseRunner`. This is the start of that.