Add typing for `BaseRunner.run_with_hooks`
As part of this I had to switch the typing of node from `GraphMemberNode`
to `ManifestNode`. We needed to do this be because `GraphMemberNode` was
_too_ wide. Specifically, `GraphMemberNode` includes the `Metric` node
which doesn't inherit from `NodeInfoMixin` and thus does not have a
`update_event_status` method. Unfortunately moving to `ManifestNode`
_excludes_ some nodes from the allowed typing that we'll probably later
on find we need to include. Maybe that means some nodes should be moved
to the `ManifestNode` definition. We might find that it would be more
appropriate however to add an `ExecutableNode` or `RunnableNode` protocol
definition, and then use that for typing.