Add type hinting to `error_result`, `ephemeral_result`, and `from_run_result` of `BaseRunner`
There are two type errors in this commit. The type hinting of `RunResult.node` and the `node`
argument of `_build_run_result` are _different_! I tried to fix this in a previous commit
by going from `GraphMemberNode` too `ManifestNode` as `GraphMemberNode` was too broad. It seems
`ManifestNode` is too narrow unfortunately. Checking the typing of `RunResult.node`, which I
should have done sooner, it uses `ResultNode`, which makes sense.

The second type error is that the type hinting of `RunResult.status` and the `status` arguement
of `_build_run_result` are _different_. The correct typing we should be that of `RunResult.status`.
Both of these typing issues will be resolved in the next commit.