Add typing to `node` property of `ExecutionContext`
We did this because in `BaseRunner.compile_and_execute` we weren't
getting type completion for accessing `ctx.node.node_info`. Said another
way, we didn't have the type context in `compile_and_execute` to know
if a `node_info` property existed on the `node` of the `ctx`. By improving
type hinting for `ExecutionContext` we were able to resolve this.