Add type hinting to exception handling methods of `BaseRunner`
Specifically methods:
- `_handle_catchable_exception`
- `_handle_internal_exception`
- `_handle_generic_exception`
- `handle_exception`

Additionally we are defining "catchable errors" twice now. Once as a union of
types in `_handle_catchable_exception` and as a tuple of classes in `handle_exception`.
We should look at refactoring this into a single definition sometime, as a change to
one should 1-1 correspond to a change in the other.